+++ go.mod
@@ -5,10 +5,9 @@ go 1.22.1
 require (
 	github.com/aws/aws-sdk-go-v2/config v1.27.18
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.55.1
-	github.com/aws/aws-sdk-go-v2/service/sts v1.28.12
 	github.com/aws/aws-sdk-go-v2/service/securityhub v1.49.2
+	github.com/aws/aws-sdk-go-v2/service/sts v1.28.12
 )
-
 
 require (
 	github.com/aws/aws-sdk-go v1.53.20
